backend/requirements.txt
Flask;
flask_sqlalchemy;
jinja2;
flask_jwt_extended;
werkzeug;
redis;
Flask-Caching;
flask-cors;
celery;
Flask-Mail;
celery[redis];